requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
whitenoise==6.11.0
dj-database-url==1.0.0
psycopg2-binary==2.9.9
python-dotenv==1.2.1
